requirements.txt
aiohttp>=3.11.13
beautifulsoup4>=4.9.1
lxml>=4.2.1
pandas>=0.25
requests>=2.22.0
urllib3>=1.25.8
html5lib>=1.0.1
tqdm>=4.43.0
xlrd>=1.2.0
openpyxl>=3.0.3
jsonpath>=0.82
tabulate>=0.8.6
decorator>=4.4.2
mini-racer>=0.12.4
nest_asyncio>=1.6.0

trafilatura
pgvector
selenium
psycopg2
openai
faiss-cpu
sentence-transformers
langchain
duckduckgo-search
alembic
mplfinance
lxml_html_clean
sqlalchemy_schemadisplay
